IGNITE-22874 Code review fixes.
