+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/exec/TableScan.java
@@ -258,7 +258,7 @@ public class TableScan<Row> implements Iterable<Row>, AutoCloseable {
                     );
 
                     if (!F.isEmpty(txChanges.get1())) {
-                        cur = new IndexScan.FilteredCursor<>(cur, txChanges.get1(), CacheSearchRow::key);
+                        cur = new FilteredCursor<>(cur, txChanges.get1(), CacheSearchRow::key);
 
                         txIter = txChanges.get2().iterator();
                     }
